+++ examples/serre_green_naghdi_1d/serre_green_naghdi_manufactured_reflecting_upwind.jl
@@ -1,11 +1,11 @@
 using OrdinaryDiffEqTsit5
 using DispersiveShallowWater
-using SummationByPartsOperators: MattssonNordström2004, derivative_operator
+using SummationByPartsOperators: Mattsson2017, upwind_operators
 
 ###############################################################################
 # Semidiscretization of the Serre-Green-Naghdi equations
 
-equations = SerreGreenNaghdiEquations1D(bathymetry_type = bathymetry_variable,
+equations = SerreGreenNaghdiEquations1D(bathymetry_type = bathymetry_mild_slope,
                                         gravity = 9.81)
 
 initial_condition = initial_condition_manufactured_reflecting
@@ -18,11 +18,11 @@ coordinates_max = 1.0
 N = 64
 mesh = Mesh1D(coordinates_min, coordinates_max, N)
 
+# create solver with SBP operators of accuracy order 2
 accuracy_order = 2
-D1 = derivative_operator(MattssonNordström2004();
-                         derivative_order = 1, accuracy_order,
-                         xmin = xmin(mesh), xmax = xmax(mesh), N = N)
-
+D1 = upwind_operators(Mattsson2017, derivative_order = 1, accuracy_order = accuracy_order,
+                      xmin = xmin(mesh), xmax = xmax(mesh),
+                      N = nnodes(mesh))
 solver = Solver(D1)
 
 # semidiscretization holds all the necessary data structures for the spatial discretization