+++ src/equations/hyperbolic_serre_green_naghdi_1d.jl
@@ -216,7 +216,7 @@ function initial_condition_manufactured(x, t,
 end
 
 """
-    source_terms_manufactured(q, x, t, equations::HyperbolicSerreGreenNaghdiEquations1D, mesh)
+    source_terms_manufactured(q, x, t, equations::HyperbolicSerreGreenNaghdiEquations1D)
 
 A smooth manufactured solution in combination with
 [`initial_condition_manufactured`](@ref).